consolidate test host creation
